Moving ImportMapProcessor.js to the root of the processors folder
