+++ src/builders/collectionBuilder.js
@@ -11,7 +11,7 @@ import PostCSSProcessor from '../processors/PostCSSProcessor.js'
 import { getFolderFilesRecursively } from '../utils/FileUtils.js'
 import { error, fatal, logChildItem, logSpacer, logTitleItem, warn } from '../utils/LoggerUtils.js'
 import { ChangeType } from '../utils/Watcher.js'
-import ImportMapProcessor from '../processors/javascript/ImportMapProcessor.js'
+import ImportMapProcessor from '../processors/ImportMapProcessor.js'
 import { plural } from '../utils/TextUtils.js'
 import ComponentBuild from '../models/ComponentBuild.js'
 import SvgProcessor from '../processors/SvgProcessor.js'
