+++ src/processors/ImportMapProcessor.js
@@ -2,9 +2,9 @@ import path, { join } from 'node:path'
 import glob from 'fast-glob'
 import picomatch from 'picomatch'
 import { init, parse } from 'es-module-lexer'
-import { exists, getFileContents, getJsonFileContents } from '../../utils/FileUtils.js'
-import { isUrl } from '../../utils/WebUtils.js'
-import FileMissingError from '../../errors/FileMissingError.js'
+import { exists, getFileContents, getJsonFileContents } from '../utils/FileUtils.js'
+import { isUrl } from '../utils/WebUtils.js'
+import FileMissingError from '../errors/FileMissingError.js'
 
 class ImportMapProcessor {
   static ImportMapFile = 'importmap.json'
